+++ include/ipc.h
@@ -8,11 +8,12 @@
 
 #include <l4/ipc.h>
 #include <types.h>
+#include <thread.h>
 
-void sys_ipc(uint64_t *param1);
+void sys_ipc(uint32_t *param1);
 uint32_t ipc_deliver(void *data);
 
-uint64_t ipc_read_mr(tcb_t *from, int i);
-void ipc_write_mr(tcb_t *to, int i, uint64_t data);
+uint32_t ipc_read_mr(tcb_t *from, int i);
+void ipc_write_mr(tcb_t *to, int i, uint32_t data);
 
 #endif /* IPC_H_ */
